separate accept bid dialog to separate component
