+++ packages/atlas/src/components/_nft/NftDialogs/AcceptBidDialog/AcceptBidList.styles.ts
