+++ packages/atlas/src/components/_nft/NftDialogs/AcceptBidDialog/index.ts
@@ -1,0 +1,1 @@
+export * from './AcceptDialog'
