+++ packages/atlas/src/components/_nft/NftDialogs/index.ts
@@ -1,0 +1,1 @@
+export * from './AllNftDialogs'
